refactor(evaluator): extract scorer and types modules (#677)
Split src/evaluator-core.ts (1107 -> ~30-line facade) by responsibility:
- evaluator-types.ts: CategoryResult/EvaluationReport/FixPlanItem/
  FixResult interfaces
- evaluator-scorers.ts: keyword tables, body helpers, and the content
  scorers (structure/description/prompt-engineering/context-efficiency)
- evaluator-scorers-policy.ts: safety/testability/license/naming scorers
ROOT_README_SUGGESTION moves to the scorers module (its only consumer);
facade re-exports the full prior surface with type-only re-exports for
interfaces. Shared test fixture strings move to evaluator-test-fixtures.
